CODING + CONTROLLER_hal Awal

+++ database/migrations/2020_11_12_111223_create_barang_models_table.php
@@ -16,7 +16,7 @@ class CreateBarangModelsTable extends Migration
         Schema::create('barang', function (Blueprint $table) {
             $table->id();
             $table->string('nama',50);
-            $table->mediumInteger('harga');
+            $table->bigInteger('harga');
             $table->smallInteger('stok');
             $table->string('gambar',255);
             $table->smallInteger('status');
